Добавлены тесты новой функциональности
